requirements.txt
numpy==1.25.0
opencv_python==4.9.0.80
pandas==2.0.3
Pillow==10.0.0
Pillow==10.2.0
scikit_learn==1.3.1
torch==2.1.2+cu121
torchvision==0.16.2+cu121
torchvision==0.16.2
transformers==4.36.1
~orch==2.0.1
~orch==2.1.2
